support symlink on decompress
